multiplication of two integers

+++ 0x04-more_functions_nested_loops/2-mul.c
@@ -1,13 +1,13 @@
 #include "holberton.h"
 /**
- * _mul - return de value of multiplication between a and b
+ * mul - return de value of multiplication between a and b
  * @a: input one
  * @b: input two
  * Return: the respective value of multiplication
  */
 int mul(int a, int b)
 {
-	int mult;
+	long int  mult;
 
 	mult = a * b;
